requirements.txt
PyMuPDF==1.26.1
python-dotenv==1.0.0
openai==1.40.0
anthropic==0.34.0
google-generativeai==0.7.2
pathlib2==2.3.7.post1